+++ TradingEdge.Database/sql/schema/tables/daily_prices.sql
@@ -6,7 +6,14 @@ CREATE TABLE IF NOT EXISTS daily_prices (
     high DOUBLE NOT NULL,
     low DOUBLE NOT NULL,
     close DOUBLE NOT NULL,
-    volume BIGINT NOT NULL,
+    -- ⚠ DOUBLE, not BIGINT. Polygon's flat files began reporting FRACTIONAL
+    -- volume on 2026-02-23 (~85% of rows from that day on; 0% before), because
+    -- some TRF prints carry fractional share counts. A BIGINT column silently
+    -- FLOORED them — e.g. 2026-08-07 ticker A: 1,525,223.372108 stored as
+    -- 1,525,223. Small in itself, but it is the same class of defect that made
+    -- split_adjusted_prices.adj_volume truncate to 0 and silently delete 436
+    -- universe ticker-days (see docs/price_adjustment.md §8).
+    volume DOUBLE NOT NULL,
     transactions BIGINT NOT NULL,
     PRIMARY KEY(ticker, date)
 );
